Revise and polish the embedding coll doc
Add a section to the `hugectr_layer_book.md` for
the embedding collection and plan file information.

Revise the notebook to refer to the API information
in the preceding file. The notebook demonstrates the
code well and the API information is updated to include
a link to the notebook.

Flake8-nb linting and black formatting for the notebook.